requirements.txt
fastapi
redis
types-redis
uvicorn
openai
python-dotenv
